move disclose-version
