+++ packages/svelte/tests/snapshot/samples/function-prop-no-getter/_expected/client/index.svelte.js
@@ -1,6 +1,6 @@
 // index.svelte (Svelte VERSION)
 // Note: compiler output will change before 5.0 is released!
-import "svelte/internal/disclose-version";
+import "svelte/internal/common/disclose-version";
 import * as $ from "svelte/internal";
 
 var Button_default = $.template(` `, true);
