+++ packages/svelte/tests/snapshot/samples/hello-world/_expected/client/index.svelte.js
@@ -1,6 +1,6 @@
 // index.svelte (Svelte VERSION)
 // Note: compiler output will change before 5.0 is released!
-import "svelte/internal/disclose-version";
+import "svelte/internal/common/disclose-version";
 import * as $ from "svelte/internal";
 
 var frag = $.template(`<h1>hello world</h1>`);
